Split HLSLocalizer.swift to 3 files
